# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin]
      summary: List all bookings (admin)
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/bookings/{id}:
    delete:
      operationId: adminCancelBooking
      tags: [admin]
      summary: Cancel a booking (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cancelled
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: adminUpdateBooking
      tags: [admin]
      summary: Update a booking's customer details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/bookings/{id}/reschedule:
    post:
      operationId: adminRescheduleBooking
      tags: [admin]
      summary: Reschedule a booking (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBookingRequest"
      responses:
        "200":
          description: Rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Slot conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/blocked-dates:
    get:
      operationId: adminListBlockedDates
      tags: [admin]
      summary: List blocked (closed) dates
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedDate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: adminAddBlockedDate
      tags: [admin]
      summary: Block a date so customers cannot book it
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockedDateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedDate"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Already blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/blocked-dates/{date}:
    delete:
      operationId: adminUnblockDate
      tags: [admin]
      summary: Remove a blocked date (re-open it)
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            description: YYYY-MM-DD
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unblocked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/service-rules:
    get:
      operationId: adminListServiceRules
      tags: [admin]
      summary: List all service-day rules with their slots
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceDayRule"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: adminCreateServiceRule
      tags: [admin]
      summary: Create a new (service, day-of-week) rule with optional initial slots
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRuleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDayRule"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A rule for that service+day already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/service-rules/{id}:
    patch:
      operationId: adminUpdateServiceRule
      tags: [admin]
      summary: Toggle whole_day_lock or active flag on a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRuleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDayRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: adminDeleteServiceRule
      tags: [admin]
      summary: Delete a service-day rule (cascades to its slots)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/service-rules/{id}/slots:
    post:
      operationId: adminAddRuleSlot
      tags: [admin]
      summary: Add a time slot (HH:MM) to a service-day rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRuleSlotRequest"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleSlot"
        "409":
          description: Slot already exists for this rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/service-rules/{id}/slots/{slotId}:
    delete:
      operationId: adminRemoveRuleSlot
      tags: [admin]
      summary: Remove a time slot from a service-day rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: slotId
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UpdateBookingRequest:
      type: object
      description: At least one field must be provided.
      minProperties: 1
      properties:
        customerName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
        vehicle:
          type: string
          minLength: 1
        notes:
          type: string
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        servicePriceCents:
          type: integer
        serviceDurationMinutes:
          type: integer
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        vehicle:
          type: string
        notes:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
        manageToken:
          type: string
          nullable: true
          description: Returned to the booking creator and via the manage endpoints. Used in the customer self-manage URL.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - serviceName
        - servicePriceCents
        - serviceDurationMinutes
        - customerName
        - email
        - phone
        - vehicle
        - notes
        - scheduledAt
        - status
        - createdAt